LayoutTests/fast/text/bidi-unprefix.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="test1" style="unicode-bidi: -webkit-isolate"></div>
<div id="test2" style="unicode-bidi: -webkit-isolate-override"></div>
<div id="test3" style="unicode-bidi: -webkit-plaintext"></div>
<div id="test4" style="unicode-bidi: isolate"></div>
<div id="test5" style="unicode-bidi: isolate-override"></div>
<div id="test6" style="unicode-bidi: plaintext"></div>
<script>
description("This test makes sure that getComputedStyle() returns the unprefixed bidi properties.");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test1')).getPropertyValue('unicode-bidi')", "isolate");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test2')).getPropertyValue('unicode-bidi')", "isolate-override");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test3')).getPropertyValue('unicode-bidi')", "plaintext");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test4')).getPropertyValue('unicode-bidi')", "isolate");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test5')).getPropertyValue('unicode-bidi')", "isolate-override");
shouldBeEqualToString("window.getComputedStyle(document.getElementById('test6')).getPropertyValue('unicode-bidi')", "plaintext");
</script>
<script src="../../resources/js-test-pre.js"></script>
</body>
</html>
